requirements.txt
pyyaml
azure-storage-blob==1.5.0